requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorclass==2.2.2
cycler==0.11.0
docopt==0.6.2
filelock==3.6.0
fonttools==4.37.1
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.22.2
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0+cu113
tqdm==4.63.0
transformers==4.16.2
typing_extensions==4.1.1
urllib3==1.26.8
